+++ src/torchjd/scalarization/__init__.py
@@ -20,10 +20,10 @@ tensor(2.)
 """
 
 from ._constant import Constant
-from ._gmean import GMean
+from ._geometric_mean import GeometricMean
 from ._mean import Mean
 from ._random import Random
 from ._scalarizer_base import Scalarizer
 from ._sum import Sum
 
-__all__ = ["Constant", "GMean", "Mean", "Random", "Scalarizer", "Sum"]
+__all__ = ["Constant", "GeometricMean", "Mean", "Random", "Scalarizer", "Sum"]